+++ src/utils.js
@@ -35,6 +35,7 @@ import {
   DCB_HOLDINGS_RECORD_ID,
   DCB_USER,
   SLIPS_TYPE,
+  REQUEST_ERROR_MESSAGE_TRANSLATION_KEYS,
 } from './constants';
 
 import css from './requests.css';
@@ -434,6 +435,17 @@ export const isSubmittingButtonDisabled = (pristine, submitting) => {
 
 export const isFormEditing = (request) => {
   return !!get(request, 'id');
+};
+
+export const getRequestErrorMessage = (error, intl) => {
+  const {
+    code = '',
+    message = '',
+  } = error;
+
+  return code && REQUEST_ERROR_MESSAGE_TRANSLATION_KEYS[code]
+    ? intl.formatMessage({ id: REQUEST_ERROR_MESSAGE_TRANSLATION_KEYS[code] })
+    : message;
 };
 
 export function resetFieldState(form, fieldName) {